Undo generalization to all builtins of PropagateBuiltins pass (now called `PropagateDeref`)
