+++ src/functional/iterator/transforms/pass_manager.py
@@ -10,7 +10,7 @@ from functional.iterator.transforms.inline_lambdas import InlineLambdas
 from functional.iterator.transforms.inline_lifts import InlineLifts
 from functional.iterator.transforms.merge_let import MergeLet
 from functional.iterator.transforms.normalize_shifts import NormalizeShifts
-from functional.iterator.transforms.propagate_builtins import PropagateBuiltins
+from functional.iterator.transforms.propagate_deref import PropagateDeref
 from functional.iterator.transforms.unroll_reduce import UnrollReduce
 
 
@@ -45,7 +45,7 @@ def apply_common_transforms(
     ir = MergeLet().visit(ir)
     ir = InlineFundefs().visit(ir)
     ir = PruneUnreferencedFundefs().visit(ir)
-    ir = PropagateBuiltins.apply(ir)
+    ir = PropagateDeref.apply(ir)
     ir = NormalizeShifts().visit(ir)
     if lift_mode != LiftMode.FORCE_TEMPORARIES:
         for _ in range(10):
