+++ tests/functional_tests/iterator_tests/test_propagate_deref.py
@@ -1,9 +1,8 @@
 from functional.ffront import itir_makers as im
-from functional.iterator import ir
-from functional.iterator.transforms.propagate_builtins import PropagateBuiltins
+from functional.iterator.transforms.propagate_deref import PropagateDeref
 
 
-def test_deref_proppagation():
+def test_deref_propagation():
     testee = im.deref_(
         im.call_(im.lambda__("inner_it")(im.lift_("stencil")("inner_it")))("outer_it")
     )
@@ -11,5 +10,5 @@ def test_deref_proppagation():
         "outer_it"
     )
 
-    actual = PropagateBuiltins.apply(testee)
+    actual = PropagateDeref.apply(testee)
     assert actual == expected
